requirements.txt
selenium
pandas
bs4
minio
fastparquet
apache-airflow-providers-apache-spark